Makefile: do not run depmod manually
The kernel build infrastructure already runs depmod properly in its
modules_install target, doing so in a cross-compilation friendly way. Running it
with no arguments is useless and also harmful since it breaks cross-compilation
contexts by scanning dependencies on the modules installed on the host instead
of those being built for the target.

Signed-off-by: Andrea Galbusera <gizero@gmail.com>