isotp: enable RX timeout handling in listen-only mode
As reported by Thomas Wagner here
https://github.com/hartkopp/can-isotp/issues/34
the timeout handling for data frames is not enabled when the isotp socket
is used in listen-only mode (sockopt CAN_ISOTP_LISTEN_MODE).
This mode is enabled by the isotpsniffer application which therefore
became inconsistend with the strict rx timeout rules when running the
isotp protocol in the operational mode.

This patch fixes this inconsistency by moving the return condition for the
listen-only mode behind the timeout handling code.

Reported-by: Thomas Wagner <thwa1@web.de>
Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>